requirements.txt
anyio==3.6.2
certifi==2022.12.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
python-dotenv==1.0.0
python-telegram-bot==20.2
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
